added retrieving encoded public RSA key of license template
fixed license template deleting

+++ backend/src/main/java/io/licensemanager/backend/service/LicenseTemplateService.java
@@ -3,6 +3,7 @@ package io.licensemanager.backend.service;
 import io.licensemanager.backend.configuration.setup.SUPPORTED_FIELD_TYPES;
 import io.licensemanager.backend.entity.LicenseTemplate;
 import io.licensemanager.backend.entity.User;
+import io.licensemanager.backend.repository.LicenseRepository;
 import io.licensemanager.backend.repository.LicenseTemplateRepository;
 import io.licensemanager.backend.repository.UserRepository;
 import io.licensemanager.backend.util.CryptoUtils;
@@ -29,6 +30,7 @@ public class LicenseTemplateService {
 
     private final LicenseTemplateRepository licenseTemplateRepository;
     private final UserRepository userRepository;
+    private final LicenseRepository licenseRepository;
 
     public List<LicenseTemplate> getLicenseTemplatesList() {
         return licenseTemplateRepository.findAll();
@@ -74,6 +76,11 @@ public class LicenseTemplateService {
         logger.debug("Deleting license template with id {}", templateId);
         Optional<LicenseTemplate> templateToDelete = licenseTemplateRepository.findById(templateId);
         if (templateToDelete.isPresent()) {
+            if (licenseRepository.existsByUsedTemplateIs(templateToDelete.get())) {
+                logger.error("Cannot delete license template with id {} - some license(s) uses it", templateId);
+
+                return false;
+            }
             licenseTemplateRepository.delete(templateToDelete.get());
 
             return true;
@@ -81,6 +88,19 @@ public class LicenseTemplateService {
         logger.error("Requested license template doesn't exist");
 
         return false;
+    }
+
+    @Transactional
+    public byte[] getTemplatePublicKey(final Long templateId) {
+        logger.debug("Getting public key of license template");
+        Optional<LicenseTemplate> template = licenseTemplateRepository.findById(templateId);
+
+        if(template.isPresent()) {
+            return template.get().getPublicKey().getEncoded();
+        }
+        logger.error("Requested license template doesn't exist");
+
+        return new byte[0];
     }
 
 }
